+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 * Add report page tooltips support
 * Make clickable elements have "pointer" cursor
 * Make task column draggable
-* Add legend/task labels font settings
+* Add legend/task labels/data labels font settings
+* Add toggle switch to match data labels colors with legend colors
 
 ### Code improvements
 * Update API to 5.11.0, tools to 5.5.1 and other packages"